Reverting to match 2.7.6a counts.

+++ source/ParametersSolo.h
@@ -91,8 +91,8 @@ public:
     
     //filtering
     char QSbase,QSmax;//quality score base and cutoff
-    double cbMinP;//for CBs with non-exact matching to WL, min posterior probability
-    //float cbMinP;//for CBs with non-exact matching to WL, min posterior probability
+    //double cbMinP;//for CBs with non-exact matching to WL, min posterior probability
+    float cbMinP;//for CBs with non-exact matching to WL, min posterior probability
 
     //cell filtering
     struct {
